digitizer: make the readout time cap relative; trigger window bounds float64
The nuclear-decay readout cap (drop deposits later than 1e5 ns) was applied to
absolute time, which drops an entire supernova interaction sitting at t0 ~ ms-s.
Make it relative to the earliest deposit — it exists to trim the late nuclear
tail (n-capture, de-excitation) after the prompt light, so relative is also the
physically correct reference. Unchanged for single-vertex/pile-up (t_ref ~ t0).
Also store trigger per_window bounds as float64 so burst-scale window times
aren't quantized to ~8 ns by a float32 cast.

+++ lucid/simulation/trigger.py
@@ -170,8 +170,8 @@ def apply_trigger(sensor_digits: dict, hits_sparse: dict, seg_hits: dict,
     counts = np.bincount(win_sorted, minlength=gates.shape[0])
     digit_offsets = np.concatenate([[0], np.cumsum(counts)]).astype(np.int32)
     per_window = {
-        "window_start": gates[:, 0].astype(np.float32),
-        "window_end": gates[:, 1].astype(np.float32),
+        "window_start": gates[:, 0].astype(np.float64),
+        "window_end": gates[:, 1].astype(np.float64),
         "digit_offsets": digit_offsets,
     }
     return new_sd, new_hits, new_seg, per_window
